rename: components 폴더 추가
