+++ src/app/mypage/alert-requests/[alertRequestId]/components/Content.tsx
