+++ src/app/mypage/alert-requests/[alertRequestId]/components/EventCard.tsx
