+++ src/app/mypage/alert-requests/[alertRequestId]/components/section/GuidelineSection.tsx
