+++ src/app/mypage/alert-requests/[alertRequestId]/components/section/ShuttleRouteInfoSection.tsx
