+++ src/app/mypage/alert-requests/[alertRequestId]/components/section/shuttle-progress-section/ShuttleProgressSection.tsx
